use imap
The pool needed to use imap instead of map. When I added some
additional logging I noticed that the entire file was being run
through before the workers began to work. This is because map
was quietly turning our generator into a list before starting to
give the workers something to do. Which defeated the purpose of
the generator in the first place!